Refactor admin UI to TypeScript React with improved architecture
