+++ packages/dwn-server/package.json
@@ -92,14 +92,15 @@
     "typescript": "^5.1.6"
   },
   "scripts": {
+    "build:admin-ui": "cd src/admin-ui && node build.js",
     "build:esm": "pnpm clean & tsc",
     "build:cjs": "pnpm build:esm && node build/create-cjs-bundle.cjs && echo '{\"type\": \"commonjs\"}' > ./dist/cjs/package.json",
-    "build": "pnpm clean && pnpm build:esm && pnpm build:cjs",
+    "build": "pnpm clean && pnpm build:esm && pnpm build:admin-ui && pnpm build:cjs",
     "clean": "rimraf dist && rimraf generated/*",
     "lint": "eslint . --ext .ts --max-warnings 0",
     "lint:fix": "eslint . --ext .ts --fix",
     "test": "pnpm build:esm && cp -R tests/fixtures dist/esm/tests && c8 mocha",
-    "server": "pnpm build:esm && node dist/esm/src/main.js"
+    "server": "pnpm build:esm && pnpm build:admin-ui && node dist/esm/src/main.js"
   },
   "husky": {
     "hooks": {
