+++ packages/dwn-server/src/admin-ui/public/index.html
@@ -4,11 +4,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>DWN Server Admin Dashboard</title>
-  <script src="https://unpkg.com/react@18/umd/react.production.min.js"></script>
-  <script src="https://unpkg.com/react-dom@18/umd/react-dom.production.min.js"></script>
-  <script src="https://unpkg.com/@babel/standalone/babel.min.js"></script>
   <script src="https://cdn.tailwindcss.com"></script>
-  <script src="https://unpkg.com/recharts@2.5.0/umd/Recharts.js"></script>
   <style>
     body {
       margin: 0;
@@ -22,6 +18,6 @@
 </head>
 <body>
   <div id="root"></div>
-  <script type="text/babel" src="./app.js"></script>
+  <script type="module" src="./app.js"></script>
 </body>
 </html>
