+++ packages/dwn-server/src/admin-ui/src/components/TenantsList.tsx
@@ -1,0 +1,78 @@
+import React from 'react';
+import type { TenantsListProps } from '../types.js';
+
+export function TenantsList({ tenants, onSelectTenant, onRefresh }: TenantsListProps): JSX.Element {
+  return (
+    <div className="bg-white rounded-lg shadow">
+      <div className="px-6 py-4 border-b border-gray-200 flex justify-between items-center">
+        <h3 className="text-lg font-medium text-gray-900">Registered Tenants</h3>
+        <button
+          onClick={onRefresh}
+          className="text-sm text-indigo-600 hover:text-indigo-800"
+        >
+          Refresh
+        </button>
+      </div>
+      
+      <div className="overflow-x-auto">
+        <table className="min-w-full divide-y divide-gray-200">
+          <thead className="bg-gray-50">
+            <tr>
+              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                DID
+              </th>
+              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Status
+              </th>
+              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Terms of Service Hash
+              </th>
+              <th className="px-6 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Actions
+              </th>
+            </tr>
+          </thead>
+          <tbody className="bg-white divide-y divide-gray-200">
+            {tenants.length === 0 ? (
+              <tr>
+                <td colSpan={4} className="px-6 py-4 text-center text-sm text-gray-500">
+                  No tenants registered
+                </td>
+              </tr>
+            ) : (
+              tenants.map((tenant) => (
+                <tr key={tenant.did} className="hover:bg-gray-50">
+                  <td className="px-6 py-4 whitespace-nowrap text-sm font-mono text-gray-900">
+                    {tenant.did}
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    {tenant.isActive ? (
+                      <span className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-green-100 text-green-800">
+                        Active
+                      </span>
+                    ) : (
+                      <span className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-red-100 text-red-800">
+                        Inactive
+                      </span>
+                    )}
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500 font-mono">
+                    {tenant.termsOfServiceHash ? tenant.termsOfServiceHash.substring(0, 16) + '...' : 'N/A'}
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
+                    <button
+                      onClick={() => onSelectTenant(tenant)}
+                      className="text-indigo-600 hover:text-indigo-900"
+                    >
+                      View Details
+                    </button>
+                  </td>
+                </tr>
+              ))
+            )}
+          </tbody>
+        </table>
+      </div>
+    </div>
+  );
+}
