+++ packages/dwn-server/src/admin-ui/src/components/index.ts
@@ -1,0 +1,5 @@
+export { LoginForm } from './LoginForm.js';
+export { ServerStats } from './ServerStats.js';
+export { TenantsList } from './TenantsList.js';
+export { TenantDetailsModal } from './TenantDetailsModal.js';
+export { Dashboard } from './Dashboard.js';
